+++ prefect_aws/client_parameters.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel, Field, FilePath, root_validator, validator
 
 class AwsClientParameters(BaseModel):
     """
-    Dataclass used to manage extra parameters that you can pass when you initialize
+    Model used to manage extra parameters that you can pass when you initialize
     the Client. If you want to find more information, see
     [boto3 docs](https://boto3.amazonaws.com/v1/documentation/api/latest/reference/core/session.html)
     for more info about the possible client configurations.